refactor: migrate mineralisation logic to fdm-calculator
