--- fdm-app/vitest.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config"
-import tsconfigPaths from "vite-tsconfig-paths"
-
-export default defineConfig({
-    plugins: [tsconfigPaths()],
-    test: {
-        environment: "node",
-        include: ["app/**/*.{test,spec}.{ts,tsx}"],
-        exclude: ["node_modules", "build"],
-    },
-})
