+++ fdm-calculator/src/index.ts
@@ -92,6 +92,29 @@ export type {
     NutrientAdviceInputs,
     NutrientAdviceResponse,
 } from "./nutrient-advice/types"
+export {
+    NmiApiError,
+    assessDataCompleteness,
+    buildDynaRequest,
+    buildNSupplyRequest,
+    getDyna,
+    getNSupply,
+    methodRequirements,
+    requestDyna,
+    requestNSupply,
+} from "./mineralisatie"
+export type {
+    DataCompleteness,
+    DynaComputeInput,
+    DynaDailyPoint,
+    DynaFertilizerAdvice,
+    DynaNitrogenBalance,
+    DynaResult,
+    NSupplyComputeInput,
+    NSupplyDataPoint,
+    NSupplyMethod,
+    NSupplyResult,
+} from "./mineralisatie"
 export type { NlvSupplyBySomParams } from "./other/nlv-supply-by-som"
 export { calculateNlvSupplyBySom } from "./other/nlv-supply-by-som"
 export type { WaterSupplyBySomParams } from "./other/water-supply-by-som"
